GH-2: Fix async processes
